% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_graph.R
\name{creer_graph}
\alias{creer_graph}
\title{Créer graph}
\usage{
creer_graph(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Créer graph
}
\examples{
mon_fichier <- system.file("example-data/iris_filtered.rds", package = "mon.super.package")
mon_iris <- readr::read_rds(mon_fichier)
mon_iris

creer_graph(mon_iris)
}
